+++ core/src/main/java/com/predic8/membrane/core/exchangestore/ForgetfulExchangeStore.java
@@ -102,7 +102,7 @@ public class ForgetfulExchangeStore implements ExchangeStore {
 
 	}
 
-	public AbstractExchange getExchangeById(int intParam) {
+	public AbstractExchange getExchangeById(long intParam) {
 		return null;
 	}
